% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauchyGM_Null.R
\name{Cauchy_Null_model}
\alias{Cauchy_Null_model}
\title{A Robust Genetic Model Based SNP-set Test using CauchyGM}
\usage{
Cauchy_Null_model(formula, data, out_type = "D")
}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the NULL model to be fitted. No need to specify outcome (y) in the formula.}

\item{data}{a list of data frames which contain the outcome (Y) and the covariates (Z) in the model.}

\item{out_type}{an indicator of the outcome type. 'C' for the continuous outcome and 'D' for the dichotomous outcome.}
}
\value{
an object that includes model parameters and residuals of the null model (there are no genetic effects on the outcome phenotypes).
}
\description{
A Robust Genetic Model Based SNP-set Test using CauchyGM
}
\author{
Yeonil Kim
}
