% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauchyGM.R
\name{CauchyGM}
\alias{CauchyGM}
\title{A Robust Genetic Model Based SNP-set Test using CauchyGM}
\usage{
CauchyGM(
  data,
  method = "CauchyGM",
  out_type = "D",
  int = 0.5,
  weights.beta = c(1, 1),
  Cauchy_Null,
  SKAT_Null
)
}
\arguments{
\item{data}{a list of data frames which contain the outcome (Y), the genotype (X_add) and the covariates (Z) in the model.}

\item{method}{a character value to specify a test procedure (default='CauchyGM'). 'CauchyGM' represents a robust genetic model based SNP-set test. 'CauchyGM-O' represents a integreated test procedure that combines CauchyGM with SKAT and burdent test.}

\item{out_type}{an indicator of the outcome type. 'C' for the continuous outcome and 'D' for the dichotomous outcome.}

\item{int}{a numeric value to set an interval of a equal sized grid between 0 and 1. This grid indicates genotypes (0=AA, Aa = 0 < c < 1, 1=aa)}

\item{weights.beta}{a vector of parameters of beta weights (default= c(1,1))}

\item{Cauchy_Null}{an output object of the Cauchy_Null_model function}

\item{SKAT_Null}{an output object of the SKAT_Null (continuous) or SKAT_Null_Model_MomentAdjust (binary) function}
}
\value{
p-value of CauchyGM or CauchyGM-O
}
\description{
A Robust Genetic Model Based SNP-set Test using CauchyGM
}
\author{
Yeonil Kim
}
