% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effScore.R
\name{eff_Score}
\alias{eff_Score}
\title{A Robust Genetic Model Based SNP-set Test: CauchyGM}
\usage{
eff_Score(y, G, Z = NULL, null_obj, out_type = "D")
}
\arguments{
\item{y}{a numeric vector of the outcome phenotype (continuous or binary phenotype).}

\item{G}{a numeric genotype vector/matrix.}

\item{Z}{a numeric vector/matrix for covariate(s). Default=NULL.}

\item{null_obj}{an object that includes model parameters and residuals of the null model (i.e., there are no genetic effects on the outcome phenotypes). Please run CauchyGM_Null to get this object.}

\item{out_type}{an indicator of the outcome type. 'C' for the continuous outcome and 'D' for the dichotomous outcome.}
}
\value{
A list of numeric vectors/matrices that contains score statistics (U) for each subject and SNP and that include the variance of the score statistics.
}
\description{
A Robust Genetic Model Based SNP-set Test: CauchyGM
}
\author{
Yeonil Kim
}
