% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAF.R
\name{MAF}
\alias{MAF}
\title{A Robust Genetic Model Based SNP-set Test using CauchyGM}
\usage{
MAF(g)
}
\arguments{
\item{g}{a numeric genotype vector/matrix. Each genotype should be coded as 0, 1, and 2 for AA, Aa, and aa, where A is a major allele and a is a minor allele.}
}
\value{
MAF for each SNP
}
\description{
A Robust Genetic Model Based SNP-set Test using CauchyGM
}
\author{
Yeonil Kim
}
